ml/document_processing/lease_parser.py
"""Lease document ingestion with Tika + Unstructured (PR#11)"""

from unstructured.partition.auto import partition
from unstructured.partition.pdf import partition_pdf
import pytesseract
from PIL import Image
from typing import Dict, List

class LeaseParser:
    """Parse lease documents and extract structured data"""
    
    def parse_lease(self, document_path: str) -> Dict:
        """Parse lease document
        
        Supports: PDF, images, scanned documents
        
        Returns:
            {
                "property_address": "123 Main St",
                "tenant_name": "John Doe",
                "start_date": "2024-01-01",
                "end_date": "2024-12-31",
                "monthly_rent": 2500,
                "security_deposit": 5000,
                "clauses": [...],
                "full_text": "..."
            }
        """
        # Use Unstructured for document parsing
        elements = partition(document_path)
        
        # Extract structured data
        extracted_data = self._extract_lease_fields(elements)
        
        # Use Tika for additional metadata
        metadata = self._extract_tika_metadata(document_path)
        
        return {**extracted_data, **metadata}
    
    def parse_scanned_lease(self, image_path: str) -> Dict:
        """Parse scanned lease with OCR
        
        Uses Tesseract OCR for text extraction from images
        """
        # OCR with Tesseract
        image = Image.open(image_path)
        text = pytesseract.image_to_string(image)
        
        # Extract fields from OCR text
        return self._extract_lease_fields_from_text(text)
    
    def _extract_lease_fields(self, elements: List) -> Dict:
        """Extract structured fields from parsed elements"""
        # NLP-based field extraction
        # Use regex + NER for entity extraction
        pass
    
    def _extract_tika_metadata(self, path: str) -> Dict:
        """Extract metadata using Apache Tika"""
        # from tika import parser
        # parsed = parser.from_file(path)
        # return parsed['metadata']
        pass
    
    def _extract_lease_fields_from_text(self, text: str) -> Dict:
        """Extract fields from plain text"""
        pass

class TenantRosterBuilder:
    """Build tenant roster from lease documents"""
    
    def create_stacking_plan(self, leases: List[Dict]) -> Dict:
        """Create stacking plan (floor → suite → tenant)
        
        Returns:
            {
                "floors": [
                    {
                        "floor_number": 1,
                        "suites": [
                            {
                                "suite_number": "101",
                                "tenant": "Acme Corp",
                                "sqft": 2500,
                                "lease_expiration": "2025-12-31"
                            }
                        ]
                    }
                ]
            }
        """
        pass
