ml/requirements.txt
# Owner deduplication (PR#7)
splink==3.9.10
python-Levenshtein==0.23.0
# ML Explainability (PR#8)
shap==0.44.0
dice-ml==0.11
# Document processing (PR#11)
unstructured==0.11.2
pytesseract==0.3.10
pillow==10.1.0
tika==2.6.0
